fixing minCalc / attempt 3

+++ sql/products/create.sql
@@ -8,12 +8,12 @@ CREATE TABLE products(
     default_max integer not null,
     default_min integer not null,
     default_date_rule varchar(256) not null,
-    default_mon_multiple integer default 1 not null,
-    default_tue_multiple integer default 1 not null,
-    default_wed_multiple integer default 1 not null,
-    default_thu_multiple integer default 1 not null,
-    default_fri_multiple integer default 1 not null,
-    default_sat_multiple integer default 1 not null,
-    default_sun_multiple integer default 1 not null,
-    default_usage integer default 1 not null
+    default_mon_multiple numeric default 1 not null,
+    default_tue_multiple numeric default 1 not null,
+    default_wed_multiple numeric default 1 not null,
+    default_thu_multiple numeric default 1 not null,
+    default_fri_multiple numeric default 1 not null,
+    default_sat_multiple numeric default 1 not null,
+    default_sun_multiple numeric default 1 not null,
+    default_usage numeric default 1 not null
 );
